fixing epsilons

+++ src/test/java/com/rb/nonbiz/testmatchers/RBMapMatchers.java
@@ -44,7 +44,7 @@ public class RBMapMatchers {
           return false;
         }
         for (K key : expected.keySet()) {
-          if (!expected.getOrThrow(key).almostEquals(actual.getOrThrow(key), epsilon.doubleValue())) {
+          if (!expected.getOrThrow(key).almostEquals(actual.getOrThrow(key), epsilon)) {
             return false;
           }
         }
@@ -67,7 +67,7 @@ public class RBMapMatchers {
           return false;
         }
         for (K key : expected.keySet()) {
-          if (!expected.getOrThrow(key).almostEquals(actual.getOrThrow(key), epsilon.doubleValue())) {
+          if (!expected.getOrThrow(key).almostEquals(actual.getOrThrow(key), epsilon)) {
             return false;
           }
         }
